requirements.txt
flask
flask-cors
numpy
opencv-python
joblib
scikit-learn
qiskit
qiskit-ibm-runtime
streamlit
python-dotenv